Restructure componen directory shape
