+++ components/custom/app-shell/page-avatar.tsx
@@ -4,7 +4,7 @@ import { cva } from 'class-variance-authority'
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
 
-import { Avatar, AvatarFallback, AvatarImage } from '../ui/avatar'
+import { Avatar, AvatarFallback, AvatarImage } from '../../ui/avatar'
 
 import { cn } from '@/lib/utils'
