+++ components/custom/timeline.tsx
@@ -10,7 +10,6 @@ const timeline = stack({
   borderInlineStartColor: 'slate.500',
 })
 
-/** @deprecated needs to be migrated from Panda to Tailwind */
 export function Timeline({ children }: PropsWithChildren) {
   return <ol className={timeline}>{children}</ol>
 }
